Remove symbol-map related functions from Host
