+++ Common/System/System.h
@@ -163,6 +163,8 @@ enum SystemProperty {
 
 	SYSPROP_CAN_JIT,
 
+	SYSPROP_HAS_DEBUGGER,
+
 	SYSPROP_KEYBOARD_LAYOUT,
 
 	SYSPROP_SKIP_UI,
