+++ Core/Host.h
@@ -31,9 +31,6 @@ public:
 	virtual void PollControllers() {}
 	virtual void ToggleDebugConsoleVisibility() {}
 
-	virtual bool AttemptLoadSymbolMap();
-	virtual void SaveSymbolMap() {}
-
 	virtual bool CreateDesktopShortcut(std::string argumentPath, std::string title) {return false;}
 
 	virtual void NotifyUserMessage(const std::string &message, float duration = 1.0f, u32 color = 0x00FFFFFF, const char *id = nullptr) {}
